Move GarminPlugin to sync package.
