+++ plugins/main/src/main/kotlin/app/aaps/plugins/main/di/PluginsModule.kt
@@ -2,7 +2,6 @@ package app.aaps.plugins.main.di
 
 import app.aaps.core.interfaces.iob.IobCobCalculator
 import app.aaps.core.interfaces.smsCommunicator.SmsCommunicator
-import app.aaps.plugins.main.general.garmin.GarminModule
 import app.aaps.plugins.main.general.persistentNotification.DummyService
 import app.aaps.plugins.main.general.smsCommunicator.SmsCommunicatorPlugin
 import app.aaps.plugins.main.general.wear.WearFragment
@@ -24,7 +23,6 @@ import dagger.android.ContributesAndroidInjector
         ActionsModule::class,
         WearModule::class,
         OverviewModule::class,
-        GarminModule::class,
     ]
 )
