--- plugins/main/src/main/kotlin/app/aaps/plugins/main/general/garmin/GarminModule.kt
@@ -1,10 +1,0 @@
-package app.aaps.plugins.main.general.garmin
-
-import dagger.Binds
-import dagger.Module
-
-@Module
-abstract class GarminModule {
-    @Suppress("unused")
-    @Binds abstract fun bindLoopHub(loopHub: LoopHubImpl): LoopHub
-}
