+++ plugins/sync/src/main/kotlin/app/aaps/plugins/sync/garmin/DeltaVarEncodedList.kt
@@ -1,4 +1,4 @@
-package app.aaps.plugins.main.general.garmin
+package app.aaps.plugins.sync.garmin
 
 import java.nio.ByteBuffer
 import java.nio.ByteOrder
