+++ plugins/sync/src/main/kotlin/app/aaps/plugins/sync/garmin/HttpServer.kt
@@ -1,4 +1,4 @@
-package app.aaps.plugins.main.general.garmin
+package app.aaps.plugins.sync.garmin
 
 import android.os.StrictMode
 import androidx.annotation.VisibleForTesting
