+++ plugins/sync/src/main/res/xml/pref_garmin.xml
