+++ plugins/sync/src/test/kotlin/app/aaps/plugins/sync/garmin/DeltaVarEncodedListTest.kt
@@ -1,4 +1,4 @@
-package app.aaps.plugins.main.general.garmin
+package app.aaps.plugins.sync.garmin
 
 
 import org.junit.jupiter.api.Assertions.assertArrayEquals
