+++ plugins/sync/src/test/kotlin/app/aaps/plugins/sync/garmin/LoopHubTest.kt
@@ -1,4 +1,4 @@
-package app.aaps.plugins.main.general.garmin
+package app.aaps.plugins.sync.garmin
 
 
 import app.aaps.core.interfaces.aps.APSResult
